@@ -33,25 +33,24 @@
     "e2e:ci": "npm run webdriver:update && npm run stub:annotations && ng e2e --webdriverUpdate=false --protractor-config 'e2e/ci.protractor.conf.js'"
   },
   "dependencies": {
-    "@angular/animations": "^7.2.15",
+    "@angular/animations": "^8.2.14",
     "@angular/cdk": "^7.3.7",
-    "@angular/cli": "^7.3.9",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
-    "@angular/http": "^7.2.15",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/platform-server": "^7.2.15",
-    "@angular/router": "^7.2.15",
+    "@angular/cli": "^8.3.29",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/platform-server": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@hmcts/frontend": "^0.0.39-alpha",
-    "@ngrx/effects": "^7.4.0",
-    "@ngrx/router-store": "^7.4.0",
-    "@ngrx/store": "^7.4.0",
-    "@ngrx/store-devtools": "^7.4.0",
-    "@nguniversal/express-engine": "^7.1.1",
-    "@nguniversal/module-map-ngfactory-loader": "^7.1.1",
+    "@ngrx/effects": "8.6.1",
+    "@ngrx/router-store": "8.6.1",
+    "@ngrx/store": "8.6.1",
+    "@ngrx/store-devtools": "8.6.1",
+    "@nguniversal/express-engine": "8.2.6",
+    "@nguniversal/module-map-ngfactory-loader": "8.2.6",
     "body-parser": "^1.19.0",
     "core-js": "^2.6.5",
     "exception-formatter": "^2.1.2",
@@ -60,19 +59,19 @@
     "global-tunnel-ng": "^2.7.1",
     "govuk-frontend": "^3.1.0",
     "moment-timezone": "^0.5.28",
-    "ngx-chips": "^1.9.8",
+    "ngx-chips": "2.0.2",
     "nodemon": "^1.19.0",
     "rxjs": "^6.5.2",
     "rxjs-compat": "^6.5.2",
     "scss-bundle": "~2.4.0",
     "session-file-store": "^1.2.0",
-    "zone.js": "^0.11.3"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.10",
-    "@angular-devkit/build-ng-packagr": "^0.13.10",
+    "@angular-devkit/build-angular": "~0.803.29",
+    "@angular-devkit/build-ng-packagr": "~0.803.29",
     "@angular-devkit/core": "^7.3.9",
-    "@angular/compiler-cli": "^7.2.15",
+    "@angular/compiler-cli": "^8.2.14",
     "@hmcts/eslint-config": "^1.4.0",
     "@ngrx/schematics": "^8.6.0",
     "@types/chai": "3.4.34",
@@ -111,16 +110,16 @@
     "karma-jasmine": "^2.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
     "karma-phantomjs-launcher": "^1.0.4",
-    "ng-packagr": "~4.7.1",
+    "ng-packagr": "^5.4.0",
     "ngrx-store-freeze": "^0.2.4",
     "pa11y-ci": "^2.4.0",
     "protractor": "^5.4.2",
     "protractor-cucumber-framework": "^6.1.1",
     "ts-node": "^8.1.0",
-    "tsickle": "^0.35.0",
+    "tsickle": "^0.37.0",
     "tslib": "^1.9.3",
     "tslint": "^5.20.1",
-    "typescript": "^3.2.4",
+    "typescript": "3.5.3",
     "typescript-string-operations": "^1.3.1",
     "webdriver-manager": "^12.1.7"
   }
