@@ -21,7 +21,7 @@
   },
   "dependencies": {
     "angular-tree-component": "^8.5.6",
-    "@swimlane/ngx-datatable": "^16.0.2",
+    "@swimlane/ngx-datatable": "16.1.1",
     "hammerjs": "^2.0.8",
     "mutable-div": "0.0.11",
     "pdfjs-dist": "2.5.207",
@@ -29,15 +29,15 @@
     "socket.io-client": "^2.3.0"
   },
   "peerDependencies": {
-    "@angular/animations": "^7.2.15",
-    "@angular/common": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
-    "@angular/platform-browser": "^7.2.15",
-    "@ngrx/effects": "^7.4.0",
-    "@ngrx/router-store": "^7.4.0",
-    "@ngrx/store": "^7.4.0",
-    "@ngrx/store-devtools": "^7.4.0"
+    "@angular/animations": "8.2.14",
+    "@angular/common": "8.2.14",
+    "@angular/core": "8.2.14",
+    "@angular/forms": "8.2.14",
+    "@angular/platform-browser": "8.2.14",
+    "@ngrx/effects": "8.6.1",
+    "@ngrx/router-store": "8.6.1",
+    "@ngrx/store": "8.6.1",
+    "@ngrx/store-devtools": "8.6.1"
   },
   "devDependencies": {
     "@types/uuid": "^3.4.4"
