@@ -89,5 +89,5 @@ describe('CommentsNavigateComponent', () => {
 class TestHostComponent {
   annotationList = [];
 
-  @ViewChild(CommentsNavigateComponent) commentsNavigateComponent;
+  @ViewChild(CommentsNavigateComponent, { static: false }) commentsNavigateComponent;
 }
